Description pour CurseForge & GitHub
Description :
SpitOn is a simple and lightweight addon that modifies the default /spit emote in World of Warcraft. With this addon, instead of the usual "You spit on the ground," it allows you to target a player and make your character spit directly on them.

Features:
Custom Emote: When using /spit, it checks if you have a target. If a target is selected, it spits on that target. If no target is selected, it defaults to "You spit on the ground."
No Display in Game: This addon only modifies the behavior of the /spit emote and does not provide any user interface elements in the game.
Classic-Compatible: Works with both World of Warcraft Classic and Retail versions.
How to Use:
Install the addon and make sure it's enabled.
Target a player and type /spit to spit on them.
If no target is selected, it will simply spit on the ground.